+++ src/language-model.ts
@@ -23,36 +23,7 @@ const debug = createDebug('chromeai');
 
 export type ChromeAIChatModelId = 'text';
 
-export interface ChromeAIChatSettings extends Record<string, unknown> {
-  temperature?: number;
-  /**
-   * Optional. The maximum number of tokens to consider when sampling.
-   *
-   * Models use nucleus sampling or combined Top-k and nucleus sampling.
-   * Top-k sampling considers the set of topK most probable tokens.
-   * Models running with nucleus sampling don't allow topK setting.
-   */
-  topK?: number;
-
-  /**
-   * Optional. A list of unique safety settings for blocking unsafe content.
-   * @note this is not working yet
-   */
-  safetySettings?: Array<{
-    category:
-      | 'HARM_CATEGORY_HATE_SPEECH'
-      | 'HARM_CATEGORY_DANGEROUS_CONTENT'
-      | 'HARM_CATEGORY_HARASSMENT'
-      | 'HARM_CATEGORY_SEXUALLY_EXPLICIT';
-
-    threshold:
-      | 'HARM_BLOCK_THRESHOLD_UNSPECIFIED'
-      | 'BLOCK_LOW_AND_ABOVE'
-      | 'BLOCK_MEDIUM_AND_ABOVE'
-      | 'BLOCK_ONLY_HIGH'
-      | 'BLOCK_NONE';
-  }>;
-}
+export interface ChromeAIChatSettings extends ChromeAISessionOptions {}
 
 function getStringContent(
   content:
@@ -79,6 +50,9 @@ export class ChromeAIChatLanguageModel implements LanguageModelV1 {
   readonly defaultObjectGenerationMode = 'json';
   readonly modelId: ChromeAIChatModelId = 'text';
   readonly provider = 'gemini-nano';
+  readonly supportsImageUrls = false;
+  readonly supportsStructuredOutputs = false;
+
   options: ChromeAIChatSettings;
 
   constructor(
@@ -106,8 +80,13 @@ export class ChromeAIChatLanguageModel implements LanguageModelV1 {
       throw new LoadSettingError({ message: 'Built-in model not ready' });
     }
 
-    const defaultOptions = await ai.defaultTextSessionOptions();
-    this.options = { ...defaultOptions, ...this.options, ...options };
+    const defaultOptions = await ai.textModelInfo();
+    this.options = {
+      temperature: defaultOptions.defaultTemperature,
+      topK: defaultOptions.defaultTopK,
+      ...this.options,
+      ...options,
+    };
 
     this.session = await ai.createTextSession(this.options);
 
@@ -219,7 +198,7 @@ export class ChromeAIChatLanguageModel implements LanguageModelV1 {
     const session = await this.getSession();
     const message = this.formatMessages(options);
     const promptStream = session.promptStreaming(message);
-    const transformStream = new StreamAI(options.abortSignal);
+    const transformStream = new StreamAI(options);
     const stream = promptStream.pipeThrough(transformStream);
 
     return {