+++ src/stream-ai.ts
@@ -1,23 +1,35 @@
-import { LanguageModelV1StreamPart } from '@ai-sdk/provider';
+import {
+  LanguageModelV1CallOptions,
+  LanguageModelV1StreamPart,
+} from '@ai-sdk/provider';
 import createDebug from 'debug';
 import { extractJSON } from './extract-json';
 
 const debug = createDebug('chromeai');
 
+export const objectStartSequence = ' ```json\n';
+export const objectStopSequence = '\n```';
+
 export class StreamAI extends TransformStream<
   string,
   LanguageModelV1StreamPart
 > {
-  public constructor(abortSignal?: AbortSignal) {
-    let textTemp = '';
+  public constructor(options: LanguageModelV1CallOptions) {
+    let buffer = '';
+    let transforming = false;
+
+    const reset = () => {
+      buffer = '';
+      transforming = false;
+    };
+
     super({
       start: (controller) => {
-        textTemp = '';
-        if (!abortSignal) return;
-        abortSignal.addEventListener('abort', () => {
+        reset();
+        if (!options.abortSignal) return;
+        options.abortSignal.addEventListener('abort', () => {
           debug('streamText terminate by abortSignal');
           controller.terminate();
-          textTemp = '';
         });
       },
       transform: (chunk, controller) => {
@@ -32,8 +44,7 @@ export class StreamAI extends TransformStream<
           finishReason: 'stop',
           usage: { completionTokens: 0, promptTokens: 0 },
         });
-        debug('stream result:', textTemp);
-        textTemp = '';
+        debug('stream result:', buffer);
       },
     });
   }