+++ vitest.config.ts
@@ -8,7 +8,7 @@ export default defineConfig({
       // If you want a coverage reports even if your tests are failing, include the reportOnFailure option
       reportOnFailure: true,
       include: ['src/**/*.ts'],
-      exclude: ['src/polyfill/*'], // TODO@jeasonstudio: finish the polyfill tests
+      exclude: ['src/polyfill/*', 'src/global.d.ts'], // TODO@jeasonstudio: finish the polyfill tests
     },
   },
 });